Fixing TZ issue
advanceByInterval was not producing intervals with expected dates in all
cases.  I worked out the precise behavior and it immediately fell out of
my head.  If you want to know, work it out for yourself.

Regardless, the problem was caused by using non-UTC functions to
manipulate the timestamp and so it was subject to local timezone
weirdness.

Added a bunch more tests and the behavior looks good now.